🚩 컴포넌트 분리하여 video 페이지에서도 선택 가능하게 변경
