+++ src/Page/Stadium/components/StadiumHours.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../scss/stadium.module.scss';
+import styles from '../scss/selectInfoBox.module.scss';
 import { StadiumHoursData } from '../../../apis/dto/scapture.dto';
 
 import Slider from 'react-slick';
@@ -9,7 +9,7 @@ import 'slick-carousel/slick/slick-theme.css';
 interface StadiumHoursProps {
   stadiumHourList: StadiumHoursData[];
   chooseSchedule: (scheduleId: number) => void;
-  isScheduleId: number;
+  isScheduleId: number | null;
 }
 
 const StadiumHours: React.FC<StadiumHoursProps> = ({
@@ -25,9 +25,6 @@ const StadiumHours: React.FC<StadiumHoursProps> = ({
     // slidesToScroll: 1,
     arrows: false,
   };
-
-  // 기본값을 리스트의 마지막 요소(오늘)로 설정
-  // const [selectedSchedule, setSelectedShcedule] = useState<number | null>(null);
 
   // // 날짜 선택 핸들러
   const handleScheduleClick = (index: number) => {
