+++ src/Page/Stadium/components/VideoList.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from '../scss/stadium.module.scss';
+import styles from '../scss/selectInfoBox.module.scss';
 
 import { useQuery } from '@tanstack/react-query';
 import { getVideoScheduled } from '../../../apis/api/stadium.api';
@@ -9,7 +9,7 @@ import { ScheduleVideo } from '../../../apis/dto/scapture.dto';
 import noDataIcon from '../../../assets/Icon/noDataIcon.svg';
 
 interface VideoListProps {
-  scheduleId: number | undefined;
+  scheduleId: number | null;
   stadiumId: number;
   toVideo: (videoId: number) => void;
 }
@@ -19,7 +19,7 @@ const VideoList: React.FC<VideoListProps> = ({ scheduleId, toVideo }) => {
   const { data: videos } = useQuery<ScheduleVideo[]>({
     queryKey: ['videoScheduled', scheduleId],
     queryFn: () => {
-      if (scheduleId !== undefined) {
+      if (scheduleId !== null) {
         return getVideoScheduled(scheduleId);
       }
       return Promise.resolve([]);
